pyproject.toml
[tool.poetry]
name = "kyb-discord-bot"
version = "0.1.0"
description = ""
authors = ["Kwon Young Bok <kybdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
openai = "^1.56.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
